feat: recharge-card-routers

+++ src/main/java/com/api/valex/Models/Recharges.java
@@ -13,7 +13,7 @@ public class Recharges {
 
     private Timestamp timestamp;
 
-    private Long amount;
+    private float amount;
 
     @ManyToOne
     @JoinColumn(name="card_id")
@@ -35,11 +35,11 @@ public class Recharges {
         this.timestamp = timestamp;
     }
 
-    public Long getAmount() {
+    public float getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
